# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ConfirmScenarioRequest": {
        "additionalProperties": false,
        "properties": {
          "confirming_identity": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Confirming Identity",
            "type": "string"
          }
        },
        "required": [
          "confirming_identity"
        ],
        "title": "ConfirmScenarioRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SimulationProgressEvent": {
        "additionalProperties": false,
        "properties": {
          "message": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "phase": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Phase",
            "type": "string"
          },
          "progress_percent": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress Percent",
            "type": "number"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "sequence": {
            "minimum": 1.0,
            "title": "Sequence",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/SimulationStatus"
          }
        },
        "required": [
          "run_id",
          "sequence",
          "status",
          "progress_percent",
          "phase",
          "message",
          "occurred_at"
        ],
        "title": "SimulationProgressEvent",
        "type": "object"
      },
...
      "SimulationRun": {
        "additionalProperties": false,
        "properties": {
          "cancellation_requested_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancellation Requested At"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "configuration": {
            "$ref": "#/components/schemas/SimulationConfiguration"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "failure": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SimulationFailureResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SimulationResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "reused_result": {
            "default": false,
            "title": "Reused Result",
            "type": "boolean"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "runtime_ms": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runtime Ms"
          },
          "scenario_fingerprint": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Scenario Fingerprint",
            "type": "string"
          },
          "scenario_id": {
            "format": "uuid",
            "title": "Scenario Id",
            "type": "string"
          },
          "simulation_fingerprint": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Simulation Fingerprint",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/SimulationStatus"
          },
          "twin_snapshot": {
            "$ref": "#/components/schemas/TwinSnapshotReference"
          }
        },
        "required": [
          "run_id",
          "scenario_id",
          "scenario_fingerprint",
          "twin_snapshot",
          "simulation_fingerprint",
          "model_version",
          "configuration",
          "status",
          "created_at"
        ],
        "title": "SimulationRun",
        "type": "object"
      },
      "SimulationStatus": {
        "enum": [
          "QUEUED",
          "RUNNING",
          "COMPLETED",
          "FAILED",
          "CANCELLED"
        ],
        "title": "SimulationStatus",
        "type": "string"
      },
...
      "StartSimulationRequest": {
        "additionalProperties": false,
        "properties": {
          "configuration": {
            "$ref": "#/components/schemas/SimulationConfiguration"
          },
          "scenario_id": {
            "format": "uuid",
            "title": "Scenario Id",
            "type": "string"
          }
        },
        "required": [
          "scenario_id"
        ],
        "title": "StartSimulationRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Keep India's energy moving.",
    "summary": "India's Energy Resilience Command Center",
    "title": "Sanjiv API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/scenario-runs": {
      "post": {
        "operationId": "start_simulation_api_v1_scenario_runs_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Sanjiv-Scenario-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sanjiv-Scenario-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartSimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Simulation",
        "tags": [
          "scenario-lab"
        ]
      }
    },
    "/api/v1/scenario-runs/{run_id}": {
      "get": {
        "operationId": "simulation_status_api_v1_scenario_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Simulation Status",
        "tags": [
          "scenario-lab"
        ]
      }
    },
    "/api/v1/scenario-runs/{run_id}/cancel": {
      "post": {
        "operationId": "cancel_simulation_api_v1_scenario_runs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Sanjiv-Scenario-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sanjiv-Scenario-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Simulation",
        "tags": [
          "scenario-lab"
        ]
      }
    },
    "/api/v1/scenario-runs/{run_id}/progress": {
      "get": {
        "operationId": "simulation_progress_api_v1_scenario_runs__run_id__progress_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SimulationProgressEvent"
                  },
                  "title": "Response Simulation Progress Api V1 Scenario Runs  Run Id  Progress Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Simulation Progress",
        "tags": [
          "scenario-lab"
        ]
      }
    },
    "/api/v1/scenario-runs/{run_id}/results": {
      "get": {
        "operationId": "simulation_results_api_v1_scenario_runs__run_id__results_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Simulation Results",
        "tags": [
          "scenario-lab"
        ]
      }
    },
    "/api/v1/scenario-runs/{run_id}/timeline": {
      "get": {
        "operationId": "simulation_timeline_api_v1_scenario_runs__run_id__timeline_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TimelinePoint"
                  },
                  "title": "Response Simulation Timeline Api V1 Scenario Runs  Run Id  Timeline Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Simulation Timeline",
        "tags": [
          "scenario-lab"
        ]
      }
    },
...
    "/api/v1/scenarios/compile": {
      "post": {
        "operationId": "compile_candidate_api_v1_scenarios_compile_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Sanjiv-Scenario-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sanjiv-Scenario-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompileScenarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioCompileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compile Candidate",
        "tags": [
          "scenario-lab"
        ]
      }
    },
...
    "/api/v1/scenarios/{scenario_id}": {
      "get": {
        "operationId": "confirmed_scenario_api_v1_scenarios__scenario_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmedScenario"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirmed Scenario",
        "tags": [
          "scenario-lab"
        ]
      }
    },
    "/api/v1/scenarios/{scenario_id}/assumptions": {
      "get": {
        "operationId": "scenario_assumptions_api_v1_scenarios__scenario_id__assumptions_get",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Assumption"
                  },
                  "title": "Response Scenario Assumptions Api V1 Scenarios  Scenario Id  Assumptions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scenario Assumptions",
        "tags": [
          "scenario-lab"
        ]
      }
    },
    "/api/v1/scenarios/{scenario_id}/audit-events": {
      "get": {
        "operationId": "scenario_audit_events_api_v1_scenarios__scenario_id__audit_events_get",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditEvent"
                  },
                  "title": "Response Scenario Audit Events Api V1 Scenarios  Scenario Id  Audit Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scenario Audit Events",
        "tags": [
          "scenario-lab"
        ]
      }
    },
    "/api/v1/scenarios/{scenario_id}/confirm": {
      "post": {
        "operationId": "confirm_candidate_api_v1_scenarios__scenario_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scenario Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Sanjiv-Scenario-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sanjiv-Scenario-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmScenarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmedScenario"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Candidate",
        "tags": [
          "scenario-lab"
        ]
      }
    },
    "/api/v1/scenarios/{scenario_id}/evidence": {
      "get": {
        "operationId": "scenario_evidence_api_v1_scenarios__scenario_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EvidenceRecord"
                  },
                  "title": "Response Scenario Evidence Api V1 Scenarios  Scenario Id  Evidence Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scenario Evidence",
        "tags": [
          "scenario-lab"
        ]
      }
    },
    "/api/v1/scenarios/{scenario_id}/validate": {
      "post": {
        "operationId": "validate_candidate_api_v1_scenarios__scenario_id__validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scenario Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Sanjiv-Scenario-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sanjiv-Scenario-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioValidationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Candidate",
        "tags": [
          "scenario-lab"
        ]
      }
    },
    "/api/v1/scenarios/{scenario_id}/validation": {
      "get": {
        "operationId": "validation_result_api_v1_scenarios__scenario_id__validation_get",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioValidationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validation Result",
        "tags": [
          "scenario-lab"
        ]
      }
    },
...
  }
}